% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crops.R
\name{crops}
\alias{crops}
\title{A technique for ...}
\usage{
crops(method, beta_min, beta_max, max_iterations = Inf, ...)
}
\arguments{
\item{method}{A ...}

\item{beta_min}{A ...}

\item{beta_max}{A ...}

\item{max_iterations}{A ...}
}
\value{
An instance of an S4 class of type crops.class.
}
\description{
A technique for ...
}
\examples{
# generate some simple data
set.seed(1)
N <- 100
data.vec <- c(rnorm(N), rnorm(N, 2), rnorm(N))

# example one - calling fpop with crops using global scope
# need the fpop library
library(pacman)
p_load(fpop)
# create a function to wrap a call to fpop for use with crops
fpop.for.crops<-function(beta)
    {
       fit <- Fpop(data.vec, beta)
       end.vec <- fit$t.est
       change.vec <- end.vec[-length(end.vec)]
       start.vec <- c(1, change.vec+1)
       segs.list <- list()
       for(seg.i in seq_along(start.vec))
           {
            start <- start.vec[seg.i]
            end <- end.vec[seg.i]
            seg.data <- data.vec[start:end]
            seg.mean <- mean(seg.data)
            segs.list[[seg.i]] <- data.frame(
                                    start, end,
                                    mean=seg.mean,
                                    seg.cost=sum((seg.data-seg.mean)^2))
            }
        segs <- do.call(rbind, segs.list)
        return(list(sum(segs$seg.cost),nrow(segs),segs$end))
    }

# now use this wrapperfunction with crops
res<-crops(fpop.for.crops,0.5*log(300),2.5*log(300))
# and plot the results
plot(res)

}
