% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crops.class.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,crops.class,data.frame-method}
\alias{plot,crops.class,missing-method}
\title{Visualisation of data, costs, penatly values and changepoint locations.}
\usage{
\S4method{plot}{crops.class,data.frame}(x, y)

\S4method{plot}{crops.class,missing}(x)
}
\arguments{
\item{x}{An instance of an S4 class produced by \code{\link{crops}}.}

\item{y}{A dataframe containing the locations and values of the data points. The data will be overlayed on the changepoint locations.}
}
\value{
A ggplot object.
}
\description{
Plot methods for an S4 object returned by \code{\link{crops}}. The plot can also overlay the original data over the changepoint locaations if required.
}
\examples{
# see the crops example

# see the crops example

}
\seealso{
\code{\link{crops}}.
}
