% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crops.class.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,crops.class-method}
\title{Pretty printing for crops results}
\usage{
\S4method{print}{crops.class}(x)
}
\arguments{
\item{x}{An instance of an S4 class produced by \code{\link{crops}}.}
}
\description{
Pretty prints a summary of a crops result
}
\examples{
# see the crops example

}
\seealso{
\code{\link{crops}}.
}
