% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crops.class.R
\docType{methods}
\name{segmentations}
\alias{segmentations}
\alias{segmentations,crops.class-method}
\title{Summary of segmentations by penalty value}
\usage{
segmentations(object)
}
\arguments{
\item{object}{An instance of an S4 class produced by \code{\link{crops}}.}
}
\value{
A data frame containing the penalties, costs, penalised costs, and changepoint locations
}
\description{
Produces a summary of the segmentations for each penalty value in the form of a data frame.
}
\examples{
# see the crops example

}
\seealso{
\code{\link{crops}}.
}
