% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crops.class.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,crops.class-method}
\title{Summary of crops result}
\usage{
\S4method{summary}{crops.class}(object)
}
\arguments{
\item{object}{An instance of an S4 class produced by \code{\link{crops}}.}
}
\description{
Prints a short summary of a crops result.
}
\examples{
# see the crops example

}
\seealso{
\code{\link{crops}}.
}
