% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crops.class.R
\docType{methods}
\name{unique}
\alias{unique}
\alias{unique,crops.class-method}
\title{Remove duplicate entries from a crops result}
\usage{
\S4method{unique}{crops.class}(x)
}
\arguments{
\item{x}{An instance of an S4 class produced by \code{\link{crops}}.}
}
\value{
An instance of the S4 class type \code{crops.class}. This is the same type as produced by the \code{\link{crops}} function.
}
\description{
Removes duplicate entries from a crops result. A duplicate entry is one having the same number of changepoints as another entry.
Note that the changepoint locations and the associated penalty and cost values are not taken into consideration. The \code{unique} function can be useful
for simplifying plots and the details produced by \code{segmentations}.
}
