% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crops.class.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,crops.class,data.frame-method}
\alias{plot,crops.class,missing-method}
\title{Visualisation of data, costs, penalty values and changepoint locations.}
\usage{
\S4method{plot}{crops.class,data.frame}(x, y)

\S4method{plot}{crops.class,missing}(x)
}
\arguments{
\item{x}{An instance of an S4 class produced by \code{\link{crops}}.}

\item{y}{A dataframe containing the locations and values of the data points. The data plot is plotted below, and is aligned with, the changepoint locations.}
}
\value{
A ggplot object. Note - if no changepoints are detected in the penalty interval [beta_min,beta_max], then the value returned is NULL.
}
\description{
Plot methods for an S4 object returned by \code{\link{crops}}. The plot can also be combined with the original data if required.
}
\examples{
# see the crops example

# see the crops example

}
\seealso{
\code{\link{crops}}.
}
